Rename ChatServiceMode to IrisChatMode, Move IrisMessageContentType, IrisMessageRequestDTO, IrisMessageResponseDTO to IrisMessage folder, Add IrisPendingContextDTO
